ResponseAnalyzer/src/requirements.txt
python-dotenv
pandas
googletrans==4.0.0-rc1
sentence-transformers
language_tool_python
langchain_ollama
langchain
opik
litellm
lexical-diversity
evaluate
nltk
rouge_score
tabulate
colorama
transformers
tqdm
